Huge movement today: WIP for the update indexes process, got through most of the logic, just need to implement the HalfHourIndex object handling/faiss specific stuff. Changed the MLRecord volume from unsigned 32 bit integer to a 4 byte float since I forgot that I'm trying to normalize the vector in place, but tested and all is working well :) Some other minor changes here and there.

+++ src/dukascopy/dukascopy.cpp
@@ -297,7 +297,7 @@ namespace Dukascopy {
                         ? firstDeltaSentinel
                         : spread - previousSpread,
 
-                .volume = candle.BidVolume
+                .volume = static_cast<float>(candle.BidVolume)
             };
 
             recFileStream.write(
